refactor(davinci): keep the inline root-hoist witness under the source budget
The 350-line ratchet failed the installment: the witness landed at 462
lines and `component.rs` crossed at 384.

- The component emit entry points — which vnode shape a component takes,
  and with it the `hoist_static_inner` position its props are decided at
  — move to `component/entry.rs` whole.
- The witness splits by subject: element roots stay in
  `davinci_s2_inline_root_hoist.rs`, component and builtin roots move to
  `davinci_s2_inline_root_hoist_components.rs`, whose gate is the
  separate disjunct rather than the shared `should_hoist` term.
- The `BindingMetadata` / `BindingTable` scaffolding every inline witness
  repeats becomes `tests/support/bindings.rs`; `davinci_s2_inline.rs`
  drops its copy too.
