+++ crates/vize_atelier_dom/tests/davinci_s2_inline_root_hoist.rs
@@ -19,8 +19,7 @@ mod support;
 
 use vize_atelier_core::options::{BindingMetadata, BindingType, CodegenMode, CodegenOptions};
 use vize_atelier_dom::DomCompilerOptions;
-use vize_s0::FxHashMap;
-use vize_s1_to_s2::{BindingKind, BindingTable, DomEmitMode, DomEmitOptions};
+use vize_s1_to_s2::{BindingTable, DomEmitMode, DomEmitOptions};
 
 const BATTERY: &[(&str, &str)] = &[
     (
@@ -47,7 +46,6 @@ const BATTERY: &[(&str, &str)] = &[
         "root_vif_child",
         r#"<div class="a"><em v-if="count">x</em></div>"#,
     ),
-    ("root_component_child", r#"<div class="a"><MyComp /></div>"#),
     (
         "two_static_props",
         r#"<div class="a" id="b" @click="handler"></div>"#,
@@ -59,14 +57,6 @@ const BATTERY: &[(&str, &str)] = &[
     (
         "root_directive_own",
         r#"<div class="a" v-show="count"></div>"#,
-    ),
-    (
-        "component_root_static",
-        r#"<MyComp class="a"><em :id="count">x</em></MyComp>"#,
-    ),
-    (
-        "component_root_empty",
-        r#"<MyComp class="a" @click="handler" />"#,
     ),
     (
         "multi_root",
@@ -104,10 +94,6 @@ const BATTERY: &[(&str, &str)] = &[
         r#"<section><div class="a"><em :id="count">x</em></div></section>"#,
     ),
     (
-        "root_component_descendant",
-        r#"<div class="a"><MyComp :id="count" /></div>"#,
-    ),
-    (
         "root_slot_descendant",
         r#"<div class="a"><em><slot /></em></div>"#,
     ),
@@ -122,18 +108,6 @@ const BATTERY: &[(&str, &str)] = &[
     (
         "root_comment_descendant",
         r#"<div class="a"><em :id="count"><!-- c --></em></div>"#,
-    ),
-    (
-        "dynamic_component_root",
-        r#"<component :is="count" class="a"><em :id="count">x</em></component>"#,
-    ),
-    (
-        "transition_root",
-        r#"<Transition class="a"><em :id="count">x</em></Transition>"#,
-    ),
-    (
-        "keepalive_root",
-        r#"<KeepAlive class="a"><em :id="count">x</em></KeepAlive>"#,
     ),
     ("root_vhtml", r#"<div class="a" v-html="msg"></div>"#),
     ("root_vtext", r#"<div class="a" v-text="msg"></div>"#),
@@ -158,18 +132,6 @@ const BATTERY: &[(&str, &str)] = &[
         r#"<div class="a"><slot name="x" /></div>"#,
     ),
     (
-        "root_named_slot_component",
-        r#"<MyComp class="a"><template #x><em :id="count">y</em></template></MyComp>"#,
-    ),
-    (
-        "root_component_text_slot",
-        r#"<MyComp class="a">text</MyComp>"#,
-    ),
-    (
-        "root_component_dyn_prop",
-        r#"<MyComp class="a" :id="count"><em :id="count">x</em></MyComp>"#,
-    ),
-    (
         "root_v_once",
         r#"<div class="a" v-once><em :id="count">x</em></div>"#,
     ),
@@ -188,59 +150,6 @@ const BATTERY: &[(&str, &str)] = &[
     (
         "root_interp_and_elem",
         r#"<div class="a">{{ count }}<em :id="count">x</em></div>"#,
-    ),
-    (
-        "comp_root_custom_dir",
-        r#"<MyComp class="a" v-focus><em :id="count">x</em></MyComp>"#,
-    ),
-    (
-        "comp_root_vslot",
-        r#"<MyComp class="a" v-slot="s"><em :id="count">{{ s }}</em></MyComp>"#,
-    ),
-    (
-        "comp_root_vif",
-        r#"<MyComp v-if="count" class="a"><em :id="count">x</em></MyComp>"#,
-    ),
-    (
-        "comp_root_vfor",
-        r#"<MyComp v-for="i in items" :key="i" class="a"><em :id="i">x</em></MyComp>"#,
-    ),
-    (
-        "comp_root_vshow",
-        r#"<MyComp class="a" v-show="count"><em :id="count">x</em></MyComp>"#,
-    ),
-    (
-        "comp_root_vmodel",
-        r#"<MyComp class="a" v-model="count"><em :id="count">x</em></MyComp>"#,
-    ),
-    (
-        "comp_root_teleport",
-        r##"<Teleport to="#a" class="b"><em :id="count">x</em></Teleport>"##,
-    ),
-    (
-        "comp_root_suspense",
-        r#"<Suspense class="a"><em :id="count">x</em></Suspense>"#,
-    ),
-    (
-        "comp_root_transition_group",
-        r#"<TransitionGroup class="a"><em :id="count" key="k">x</em></TransitionGroup>"#,
-    ),
-    (
-        "comp_root_dyn_is",
-        r#"<component is="MyComp" class="a"><em :id="count">x</em></component>"#,
-    ),
-    ("comp_root_no_children", r#"<MyComp class="a" />"#),
-    (
-        "comp_root_slot_outlet_child",
-        r#"<MyComp class="a"><slot /></MyComp>"#,
-    ),
-    (
-        "multi_root_component",
-        r#"<MyComp class="a"><em :id="count">x</em></MyComp><p class="b"><em :id="count">y</em></p>"#,
-    ),
-    (
-        "multi_root_builtin",
-        r#"<Transition class="a"><em :id="count">x</em></Transition><p class="b"><em :id="count">y</em></p>"#,
     ),
     (
         "elem_root_custom_dir_dyn",
@@ -277,7 +186,7 @@ const BATTERY: &[(&str, &str)] = &[
 ];
 
 fn metadata() -> BindingMetadata {
-    let entries: &[(&str, BindingType)] = &[
+    support::bindings::script_setup_metadata(&[
         ("count", BindingType::SetupRef),
         ("msg", BindingType::SetupLet),
         ("handler", BindingType::SetupConst),
@@ -285,49 +194,7 @@ fn metadata() -> BindingMetadata {
         ("MyComp", BindingType::SetupConst),
         ("vFocus", BindingType::SetupConst),
         ("items", BindingType::SetupConst),
-    ];
-    let mut bindings = FxHashMap::default();
-    for (name, kind) in entries {
-        bindings.insert((*name).into(), *kind);
-    }
-    BindingMetadata {
-        bindings,
-        props_aliases: FxHashMap::default(),
-        is_script_setup: true,
-    }
-}
-
-fn binding_kind(kind: BindingType) -> BindingKind {
-    match kind {
-        BindingType::SetupLet => BindingKind::SetupLet,
-        BindingType::SetupMaybeRef => BindingKind::SetupMaybeRef,
-        BindingType::SetupRef => BindingKind::SetupRef,
-        BindingType::SetupReactiveConst => BindingKind::SetupReactiveConst,
-        BindingType::SetupConst => BindingKind::SetupConst,
-        BindingType::Props => BindingKind::Props,
-        BindingType::PropsAliased => BindingKind::PropsAliased,
-        BindingType::Data => BindingKind::Data,
-        BindingType::Options => BindingKind::Options,
-        BindingType::LiteralConst => BindingKind::LiteralConst,
-        BindingType::JsGlobalUniversal => BindingKind::JsGlobalUniversal,
-        BindingType::JsGlobalBrowser => BindingKind::JsGlobalBrowser,
-        BindingType::JsGlobalNode => BindingKind::JsGlobalNode,
-        BindingType::JsGlobalDeno => BindingKind::JsGlobalDeno,
-        BindingType::JsGlobalBun => BindingKind::JsGlobalBun,
-        BindingType::VueGlobal => BindingKind::VueGlobal,
-        BindingType::ExternalModule => BindingKind::ExternalModule,
-    }
-}
-
-fn table(metadata: &BindingMetadata) -> BindingTable {
-    BindingTable::new(
-        metadata
-            .bindings
-            .iter()
-            .map(|(name, kind)| (name.as_str(), binding_kind(*kind))),
-        [],
-        metadata.is_script_setup,
-    )
+    ])
 }
 
 fn dom_options(metadata: &BindingMetadata, inline: bool) -> DomCompilerOptions {
@@ -353,7 +220,7 @@ fn emit_options<'a>(table: &'a BindingTable, inline: bool) -> DomEmitOptions<'a>
 #[test]
 fn inline_root_hoist_matches_the_shipped_dom_lane() {
     let metadata = metadata();
-    let table = table(&metadata);
+    let table = support::bindings::binding_table(&metadata);
     support::assert_s2_matches_shipped_with_options(
         BATTERY,
         &dom_options(&metadata, true),
@@ -369,7 +236,7 @@ fn inline_root_hoist_matches_the_shipped_dom_lane() {
 #[test]
 fn the_same_battery_keeps_its_root_props_inline_without_the_option() {
     let metadata = metadata();
-    let table = table(&metadata);
+    let table = support::bindings::binding_table(&metadata);
     support::assert_s2_matches_shipped_with_options(
         BATTERY,
         &dom_options(&metadata, false),
@@ -382,7 +249,7 @@ fn the_same_battery_keeps_its_root_props_inline_without_the_option() {
 #[test]
 fn the_arm_is_what_moves_the_root_props_into_the_preamble() {
     let metadata = metadata();
-    let table = table(&metadata);
+    let table = support::bindings::binding_table(&metadata);
     let allocator = vize_s0::Allocator::new();
     let emit = |src: &str, inline: bool| {
         vize_s1_to_s2::emit_dom_source_with_options(
@@ -437,26 +304,4 @@ fn the_arm_is_what_moves_the_root_props_into_the_preamble() {
     // not fire even at the root.
     let component_descendant = r#"<div class="a"><MyComp :id="count" /></div>"#;
     assert_eq!(hoist_line(component_descendant, true), "");
-
-    // A component root reaches the arm the same way its element sibling
-    // does — including the builtin helpers, whose ordinary hoist gate is
-    // reconstructed from a codegen shape the arm sits in front of.
-    assert_eq!(
-        hoist_line(r#"<MyComp class="a"><em :id="count">x</em></MyComp>"#, true),
-        "const _hoisted_1 = { class: \"a\" }"
-    );
-    assert_eq!(
-        hoist_line(
-            r#"<Transition class="a"><em :id="count">x</em></Transition>"#,
-            true
-        ),
-        "const _hoisted_1 = { class: \"a\" }"
-    );
-    assert_eq!(
-        hoist_line(
-            r#"<KeepAlive class="a"><em :id="count">x</em></KeepAlive>"#,
-            true
-        ),
-        "const _hoisted_1 = { class: \"a\" }"
-    );
 }
