+++ crates/vize_atelier_dom/tests/support/bindings.rs
@@ -1,0 +1,57 @@
+//! The binding-metadata inputs the P2-11 witnesses hand both lanes: the
+//! shipped `BindingMetadata` and the neutral `BindingTable` that mirrors
+//! it, from one `(name, kind)` list.
+
+use vize_atelier_core::options::{BindingMetadata, BindingType};
+use vize_s0::FxHashMap;
+use vize_s1_to_s2::{BindingKind, BindingTable};
+
+/// A `<script setup>` metadata block over `entries`.
+pub fn script_setup_metadata(entries: &[(&str, BindingType)]) -> BindingMetadata {
+    let mut bindings = FxHashMap::default();
+    for (name, kind) in entries {
+        bindings.insert((*name).into(), *kind);
+    }
+    BindingMetadata {
+        bindings,
+        props_aliases: FxHashMap::default(),
+        is_script_setup: true,
+    }
+}
+
+/// The S2 emitter's neutral mirror of `metadata`.
+pub fn binding_table(metadata: &BindingMetadata) -> BindingTable {
+    BindingTable::new(
+        metadata
+            .bindings
+            .iter()
+            .map(|(name, kind)| (name.as_str(), binding_kind(*kind))),
+        metadata
+            .props_aliases
+            .iter()
+            .map(|(local, key)| (local.as_str(), key.as_str())),
+        metadata.is_script_setup,
+    )
+}
+
+fn binding_kind(kind: BindingType) -> BindingKind {
+    match kind {
+        BindingType::SetupLet => BindingKind::SetupLet,
+        BindingType::SetupMaybeRef => BindingKind::SetupMaybeRef,
+        BindingType::SetupRef => BindingKind::SetupRef,
+        BindingType::SetupReactiveConst => BindingKind::SetupReactiveConst,
+        BindingType::SetupConst => BindingKind::SetupConst,
+        BindingType::Props => BindingKind::Props,
+        BindingType::PropsAliased => BindingKind::PropsAliased,
+        BindingType::Data => BindingKind::Data,
+        BindingType::Options => BindingKind::Options,
+        BindingType::LiteralConst => BindingKind::LiteralConst,
+        BindingType::JsGlobalUniversal => BindingKind::JsGlobalUniversal,
+        BindingType::JsGlobalBrowser => BindingKind::JsGlobalBrowser,
+        BindingType::JsGlobalNode => BindingKind::JsGlobalNode,
+        BindingType::JsGlobalDeno => BindingKind::JsGlobalDeno,
+        BindingType::JsGlobalBun => BindingKind::JsGlobalBun,
+        BindingType::VueGlobal => BindingKind::VueGlobal,
+        BindingType::ExternalModule => BindingKind::ExternalModule,
+    }
+}
